requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymupdf
python-docx
openpyxl
python-pptx
ezdxf
matplotlib
pywin32
xlrd
rarfile
 
